backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
openai==1.3.0
Werkzeug==2.3.7
hypercorn==0.25.0
asgiref==3.7.2 